Move config parser to Utils module
