+++ spec/backward_compatibility_specs/configuration_spec_bc.rb
@@ -6,7 +6,7 @@ describe "Webpacker::Configuration" do
   context "with standard webpacker.yml" do
     let(:config) do
       config_path = Pathname.new(File.expand_path("./webpacker_test_app/config/webpacker.yml", __dir__))
-      config_hash = Webpacker::Helper.parse_config_file_to_hash(config_path)
+      config_hash = Webpacker::Utils.parse_config_file_to_hash(config_path)
       Webpacker::Configuration.new(
         root_path: ROOT_PATH,
         custom_config: config_hash,
@@ -181,7 +181,7 @@ describe "Webpacker::Configuration" do
 
   context "with webpacker config file containing public_output_path entry" do
     config_path = Pathname.new(File.expand_path("./webpacker_test_app/config/webpacker_public_root.yml", __dir__))
-    config_hash = Webpacker::Helper.parse_config_file_to_hash(config_path)
+    config_hash = Webpacker::Utils.parse_config_file_to_hash(config_path)
     config = Webpacker::Configuration.new(
       root_path: ROOT_PATH,
       custom_config: config_hash,
@@ -196,7 +196,7 @@ describe "Webpacker::Configuration" do
 
   context "with webpacker config file containing manifext_path entry" do
     config_path = Pathname.new(File.expand_path("./webpacker_test_app/config/webpacker_manifest_path.yml", __dir__))
-    config_hash = Webpacker::Helper.parse_config_file_to_hash(config_path)
+    config_hash = Webpacker::Utils.parse_config_file_to_hash(config_path)
     config = Webpacker::Configuration.new(
       root_path: ROOT_PATH,
       custom_config: config_hash,
@@ -218,7 +218,7 @@ describe "Webpacker::Configuration" do
 
       let(:config) {
         config_path = Pathname.new(File.expand_path("./webpacker_test_app/config/webpacker_no_precompile.yml", __dir__))
-        config_hash = Webpacker::Helper.parse_config_file_to_hash(config_path)
+        config_hash = Webpacker::Utils.parse_config_file_to_hash(config_path)
         Webpacker::Configuration.new(
           root_path: ROOT_PATH,
           custom_config: config_hash,
@@ -248,7 +248,7 @@ describe "Webpacker::Configuration" do
   context "with webpacker config file with defaults fallback" do
     let(:config) do
       config_path = Pathname.new(File.expand_path("./webpacker_test_app/config/webpacker_defaults_fallback.yml", __dir__))
-      config_hash = Webpacker::Helper.parse_config_file_to_hash(config_path)
+      config_hash = Webpacker::Utils.parse_config_file_to_hash(config_path)
       Webpacker::Configuration.new(
         root_path: ROOT_PATH,
         custom_config: config_hash,
@@ -268,7 +268,7 @@ describe "Webpacker::Configuration" do
   context "falls back to bundled production config for custom environments" do
     let(:config) do
       config_path = Pathname.new(File.expand_path("./webpacker_test_app/config/webpacker_defaults_fallback.yml", __dir__))
-      config_hash = Webpacker::Helper.parse_config_file_to_hash(config_path)
+      config_hash = Webpacker::Utils.parse_config_file_to_hash(config_path)
       Webpacker::Configuration.new(
         root_path: ROOT_PATH,
         custom_config: config_hash,
