+++ spec/configuration_spec.rb
@@ -1,5 +1,5 @@
 require_relative "spec_helper_initializer"
-require "shakapacker/helper"
+require "shakapacker/utils"
 
 describe "Shakapacker::Configuration" do
   ROOT_PATH = Pathname.new(File.expand_path("test_app", __dir__))
@@ -29,7 +29,7 @@ describe "Shakapacker::Configuration" do
   context "with standard shakapacker.yml" do
     let(:config) do
       config_path = Pathname.new(File.expand_path("./test_app/config/shakapacker.yml", __dir__))
-      config_hash = Shakapacker::Helper.parse_config_file_to_hash(config_path)
+      config_hash = Shakapacker::Utils.parse_config_file_to_hash(config_path)
       Shakapacker::Configuration.new(
         root_path: ROOT_PATH,
         custom_config: config_hash,
@@ -203,7 +203,7 @@ describe "Shakapacker::Configuration" do
 
   context "with shakapacker config file containing public_output_path entry" do
     config_path = Pathname.new(File.expand_path("./test_app/config/shakapacker_public_root.yml", __dir__))
-    config_hash = Shakapacker::Helper.parse_config_file_to_hash(config_path)
+    config_hash = Shakapacker::Utils.parse_config_file_to_hash(config_path)
     config = Shakapacker::Configuration.new(
       root_path: ROOT_PATH,
       custom_config: config_hash,
@@ -218,7 +218,7 @@ describe "Shakapacker::Configuration" do
 
   context "with shakapacker config file containing manifext_path entry" do
     config_path = Pathname.new(File.expand_path("./test_app/config/shakapacker_manifest_path.yml", __dir__))
-    config_hash = Shakapacker::Helper.parse_config_file_to_hash(config_path)
+    config_hash = Shakapacker::Utils.parse_config_file_to_hash(config_path)
     config = Shakapacker::Configuration.new(
       root_path: ROOT_PATH,
       custom_config: config_hash,
@@ -239,7 +239,7 @@ describe "Shakapacker::Configuration" do
 
       let(:config) {
         config_path = Pathname.new(File.expand_path("./test_app/config/shakapacker_no_precompile.yml", __dir__))
-        config_hash = Shakapacker::Helper.parse_config_file_to_hash(config_path)
+        config_hash = Shakapacker::Utils.parse_config_file_to_hash(config_path)
         config = Shakapacker::Configuration.new(
           root_path: ROOT_PATH,
           custom_config: config_hash,
@@ -268,7 +268,7 @@ describe "Shakapacker::Configuration" do
   context "with shakapacker config file with defaults fallback" do
     let(:config) do
       config_path = Pathname.new(File.expand_path("./test_app/config/shakapacker_defaults_fallback.yml", __dir__))
-      config_hash = Shakapacker::Helper.parse_config_file_to_hash(config_path)
+      config_hash = Shakapacker::Utils.parse_config_file_to_hash(config_path)
       config = Shakapacker::Configuration.new(
         root_path: ROOT_PATH,
         custom_config: config_hash,
@@ -288,7 +288,7 @@ describe "Shakapacker::Configuration" do
   context "falls back to bundled production config for custom environments" do
     let(:config) do
       config_path = Pathname.new(File.expand_path("./test_app/config/shakapacker_defaults_fallback.yml", __dir__))
-      config_hash = Shakapacker::Helper.parse_config_file_to_hash(config_path)
+      config_hash = Shakapacker::Utils.parse_config_file_to_hash(config_path)
       Shakapacker::Configuration.new(
         root_path: ROOT_PATH,
         custom_config: config_hash,
@@ -307,7 +307,7 @@ describe "Shakapacker::Configuration" do
   context "#source_entry_path" do
     let(:config) do
       config_path = Pathname.new(File.expand_path("./test_app/config/shakapacker.yml", __dir__))
-      config_hash = Shakapacker::Helper.parse_config_file_to_hash(config_path)
+      config_hash = Shakapacker::Utils.parse_config_file_to_hash(config_path)
       Shakapacker::Configuration.new(
         root_path: ROOT_PATH,
         custom_config: config_hash,
